+++ src/editor/resources/shaders/textures/vertex.glsl
@@ -2,7 +2,7 @@
 
 layout(location=0) in vec4 position;
 layout(location=1) in vec2 texCoord;
-layout(location=2) in float layer;
+layout(location=2) in int layer;
 layout(location=3) in float T_Index;
 
 out vec2 v_TexCoord;
