+++ src/editor/resources/shaders/triangles/vertex.glsl
@@ -2,7 +2,7 @@
 
 layout(location=0) in vec4 position;
 layout(location=1) in vec4 color;
-layout(location=2) in float layer;
+layout(location=2) in int layer;
 
 out vec4 out_color;
